rework user and token context and most of its usages to use unstated-next
