+++ src/components/PageLayout/Authentication/Authentication.tsx
@@ -2,7 +2,8 @@ import axios, {AxiosError} from 'axios'
 import {FC, useState} from 'react'
 import {useCookies} from 'react-cookie'
 
-import {useSetWebstromToken, useUser} from '@/utils/UserContext'
+import {UserContainer} from '@/utils/UserContainer'
+import {WebstromTokenContainer} from '@/utils/WebstromTokenContainer'
 
 import {Overlay} from '../../Overlay/Overlay'
 import {LoginForm} from '../LoginForm/LoginForm'
@@ -12,8 +13,8 @@ export const Authentication: FC = () => {
   const [displayAuthenticationOverlay, setDisplayAuthenticationOverlay] = useState(false)
   const [displayLogin, setDisplayLogin] = useState(true) // true -> zobrazí sa login, false -> zobrazí sa registrácia
   const [, , removeCookie] = useCookies(['webstrom-token']) // ToDo: remove
-  const user = useUser()
-  const setWebstromToken = useSetWebstromToken()
+  const {user} = UserContainer.useContainer()
+  const {setWebstromToken} = WebstromTokenContainer.useContainer()
 
   const toggleDisplayAuthenticationOverlay = () => {
     setDisplayAuthenticationOverlay((prevDisplay) => {
